Significant reformatting/refactoring of REGRESSIONS
To go along with our nice, clean, new cron job identifiers, I've
refactored the REGRESSIONS file, in part inspired by the cleanup, in
part enabled by it.  This doesn't do anything like get us to a
completely happy place with REGRESSIONS, but hopefully people will
understand it better and it'll be useful for regressions week.
(And hopefully after REGRESSIONS week, it'll be much emptier).

There's still more to do here (esp. since not all tests have come
in yet under the new system), but this is a good enough start
that it's worth snapshotting and pushing out for people to see.

The major changes were:

* Updated comments at top of file to reflect current thinking/format

* Reformatted category entries to match nightly runs

* Made the implied/mental inheritance between categories explicit in
  the category header itself rather than vague entries like "along
  with the above"

* Introduced new "wildcard/abstract categories" (other than just
  "general" before) to serve as roots for common configurations
  like "gasnet*" rather than relying on "gasnet-everything" to
  be the base case for all GASNet testing (it wasn't always
  accurate in this respect)

* Tightened up the wording for the category names, reviewed by
  dates, etc. to be pithier.

* Reordered tests to reflect what seems like a more logical ordering:
  Primary environments, interesting configurations, gasnet, xe/xc,
  lower-priority environments, then distribution testing.

* Also tried to unify more on the name "sporadic" when describing
  nondeterministic cases and move these to the end of sections,
  but more work remains here.