Add CHPL_TEST_UNIQUIFY_EXE allowing parallel runs of sub_test
If CHPL_TEST_UNIQUIFY_EXE is set, the test has the PID of the sub_test
process appended to its executable name. This allows you to run multiple
start_tests over the same set of tests in parallel.

This feature is disabled by default and may fail with tests using
.prediffs and similar features of the test system.