Fix NUMA build failure.
Application of FLAG_COMPILER_GENERATED caused locale.name to be selected in favor of
NumaDomain.name.  Assignment to this in the constructor then looked like an attempt to
write to an rvalue, so the Chapel compiler flagged it as an error.

I decided to back out the change.  It seems like in general a user wants the local field
name to bind at least as tightly as an inherited getter.  The compiler generates getter
and setter functions only if they are not provided by the user, so the "compiler
generated" flag is not really needed in this case.

I also backed out the change to classes/diten/nearestMutualParentClass, as the "compiler
generated" pragma is no longer needed there.