Update record_from_string.bad file.
This is an improvement resulting from PR#667.

Because FLAG_COMPILER_GENERATED was not attached to the compiler-generated record
assignment function, the assignment type mismatch error was attributed to the first line
of the record definition (line 1) rather than the line on which it is invoked (line 11).
Now the error is attributed to the user source code line that causes it, which is at once
more accurate and more useful.