Accept improved behavior of explainIntantiation.
Making _defaultOf() call the zero-args constructor by name removed the inexplicable second
call to the range constructor in DefaultRectangularDom (a wrapper, perhaps), and replaced
it with an instantiation of the range type.  The runtime behavior is unaffected.

That instantiation chain seems more logical, so I'm putting this down as an improvement.