Remove generateConcreteConstructorCall() and helpers.
On-the-fly generation of concrete constructor calls is no longer necessary, because the
code that generates _defaultOf() was changed to call the zero-args constructor by name.
Since the insertion of the zero-args constructor call was moved there (buildDefaultArgs),
insertion at resolution time is no longer necessary.  So the above function and its
helpers have become dead code.

Also, I set up some asserts that will permit merging the tails of the two initialization
clauses.