Add pragma "compiler generated" to range.range as a workaround.
After reworking the compiler-generated version of _defaultOf() for a record type so it
calls the no-args constructor by name(!), the single test case
types/records/sungeun/chapelTypes/record_range fails because name lookup during resolution
does not work as specified.

I excerpted the pertinent text from the spec and added this to the test to underscore what
is expected.

I restored the desired behavior for this test case by adding the "compiler generated" flag
to the constructor range.range in ChapelRange.chpl, but this is an ugly hack.  The correct
solution would be to perform name lookup during resolution as required by the spec.
However, that change is likely to be significant, so I left it as a future effort.