Fix calculation of physical memory for Cygwin
It used to be the case that using sysconf() on Cygwin was a reliable
way to find out how much memory the system had (just like linux), but
that doesn't seem to be the case anymore.  For recent runs, it has
been returning a pretty small number for the number of pages
available.  Searching the web, I didn't easily find any explanation
for why this wouldn't work anymore (and plenty of guidance saying to
do it this way).

The other way to do it is to use the Windows API which seems to work
and returns the right value, so I spliced in some code to do that for
Cygwin now by default.  This makes the memory counting tests work on
my Cygwin machine now.