Turn off mallocOutOfMemory for darwin, as it does not behave appropriately
This test depends on the ability to limit the amount of virtual memory so that
the process will attempt to allocate too much and exit with an error message.
Unfortunately, various internet sources tell me that Darwin's ways to limit
virtual memory do not work.  Since the alternative is a test that would stomp
all over everything if left unchecked and that will always time out in our
testing system, I'm turning it off for darwin.