Silencing --baseline regression from last night and adding a future
It seems that using a 'ref' var (currently 'refvar') within an
iterator breaks when the iterator is represented by an object as
this test demonstrates.  When we're not compiling --baseline,
the iterator object is optimized away, so things work fine; with
--baseline, it segfaults.  I'm not 100% sure I'm diagnosing this
properly, but it makes sense and seems to match what I'm seeing
in the generated code.  See the .future file for details.