Add align.skipif to squelch weird Intel codegen error.
The Intel compiler causes the result of the modulo operator to have the "wrong" sign, but
only under very special circumstances (no -g flag, only in chpl__mod() so far).  In a
similar test, the modulo operator gives a result with the "right" sign.  The C spec does
not require that the sign of the result match the sign of the dividend (with a positive
modulus), but it seems reasonable to expect that the compiler is at least self-consistent.

So, I rate this as an Intel compiler bug and propose to just simply avoid it.