Fix TooManyThreads valgrind error
This test requires that CHPL_RT_NUM_THREADS_PER_LOCALE is unset for correct
output, but our valgrind testing sets this variable in order to run.  Therefore,
the test is not useful under these circumstances.  The effort to make the
prediff more correctly react to this variable being set seemed much greater than
the payoff, so I left a note in the test of how it could be done and turned it
off for valgrind.