Quiet align.chpl errors with Intel and PGI compilers.
The odd behavior on Intel and PGI compilers was due to how a 32-bit temporary was being
inlined:
 - On GCC, the declared size of the temporary (int32_t) was being obeyed, either by being
    written back to a 32-bit memory location, or by being sign-extended inline.
 - On Intel and PGI, the 32-bit result was stored in a 64-bit register and not 32-bit
    sign-extended.

The intent of the test was to demonstrate correct behavior with the first 64-bit integer
not representable as a 32-bit integer.  To avoid compiler dependencies, this constant is
now hard-coded.

The compiler-dependent behavior is captured in a new future with a suggestion for
specifying the behavior that is expected.