Quiet execflags/bradc/gdbddash/declint2 output on PrgEnv-pgi
The existing prediff script looks for "gdbShouldBreakHere" but in the fast tests, GDB
doesn't print it out at all.

Under PGI, the breakpoint location is printed out with a leading dot-slash.

I solved both problems by modifying the prediff to look for the runtime output
  declint2.chpl:7: error: halt reached - array index out of bounds: (0)
instead.

Since the point of the test is to show that we can run under GDB, if we get as far as
generating error output, we have run the program.  Done.