Move the failure mode of externFloat4calls.someFields.chpl to a future and disable the mode it fails here.
I copied this test to a separate file so that it could be futurized without
losing the accurate results produced under other circumstances.

Throwing --no-remove-copy-calls causes this test to lose information for extern
record fields not named within Chapel, as an autoCopy does not know to copy that
information over.  The new future always throws this flag, to ensure the failure
mode is visible.