Turn off mallocOutOfMemory for cygwin
Cygwin also fails to define a useful ulimit function, so any attempts to limit
the memory of this test will not work.  Even though the test will still run and
complete with a higher number of iterations, I think we should treat it the way
we treat darwin and skip it.