Removing Sung's lulesh studies test
With the bug fixes of the earlier commits here, I verified that the
release version subsumes Sung's version and outperforms it when
compiled in the similar configuration (probably due to using atomics
rather than syncs).  Given that this test is now redundant and that
we're getting timeouts on it in many configurations, I think it's time
to remove it.