Remove a future file, it was resolved yesterday
Yesterday, I made our domain debugging output cast its message to the leak-free
string record implementation in order to silence some baseline failures.  Turns
out this was the reason this future was still failing (though I suspect that
Mike resolved the actual problem earlier).  I made a note in the test itself
about the expected output (it still says it leaks 8 bytes but that's because a
global array is now longer than it was before, and that array will be taken care
of upon completion of the test), and removed the skipif as this test does work
with CHPL_COMM=gasnet.