Back dest==src test out of chpl_memcpy().
When -Werror=uninitialized is tossed, GCC complains about the version of chpl_memcpy()
where memcpy() is called only if dest and src differ.  It apparently does not reason that
dest is valid since it is already equal to src.  So I backed out the aggressive
conditional call to the system memcpy, and instead moved the dest != src test back to
chpl_gen_comm_get().

I also noticed that code used in building the launcher cannot see the definition of
chpl_memcpy(), so I reverted those calls to being direct calls to the system memcpy().

TODO: We may still get hung up on the assert, in which case we can address the cases
proactively rather than waiting for valgrind to flag them.