Disable task table by default
This commit disables task tracking and the task table by default.
This cuts down on the size of the generated code and compile time
significantly.  The argument is that if your program fails
deterministically enough to be able to re-run with task tracking,
you can afford the time to recompile it.

We could make this default to on for CHPL_TASKS=fifo (the only place
it works currently anyway), but that suggests that once qthreads
supported it, we'd want it to default on there too, and I'm not sure
that's the case.  Or else we'd want to disable it in testing by
default.