Fix minimal modules bug
The minimal modules bug on PrgEnv-cray was due to an empty array
initializer list with an inferred size, which is apparently illegal.
The fix was as simple as printing out a single null pointer in the
list in this case (something that we already did for a previous
list -- I simply mimicked the logic).