Attempted fix for portability of uminusVsTimesPrec test
This test has been failing on prgenv-cray due to its reliance on
signed integer wraparound.  The intent of the test is interesting to
me (trying to rationalize relative precedence between unary minus and
multiplication), and a quick promotion of it from an int(32) test to
an int(64) test seems to work for prgenv-cray, so I'd like to try
running it that way to see if it makes it work everywhere.  If it
doesn't, we can decide to either skipif the test or retire it.