Fix for last night's --baseline regressions
This commit fixes last night's glob baseline regressions.  The
cause of the problem is that, in --baseline mode, we introduce
variables of the fake glob_err_fn_t type that I declared in the
Chapel module; but since that type is never defined in C, we
get C runtime link errors.  This may be similar to the "extra
copies of extern types under baseline" issue that Lydia's been
running into.

I also found that I had to add wordexp and glob to the reserved
symbol names list even though I don't think we've had to do this
in the past (even though it makes sense that we would need to).
I wonder if this is somehow a darwin+baseline-specific thing?