Fix for Filerator valgrind regressions
I'm not 100% sure why yet, but something about my use of an
extern opaque type to represent a function pointer and
declaration of NULL for the actual to that argument caused
a valgrind error in getValType() (worth exploring further).

In the interest of closing this down quickly and cleanly,
I went back to the original implementation approach of simply
wrapping glob() in a chpl_glob() wrapper that only deals with
the three arguments we need to deal with in Chapel.  This
resolves both the valgrind and baseline issues.