Updating the .bad file here to reflect a shift in compiler numbering
This .bad file is sensitive to the numbering of compiler-named
identifiers, which shifted with the removal of the TaskTable
by default.  In the long-run, we may want to look at a more
maintainable way of supporting this .bad, or to just remove it,
but for now I'll keep it going.