Update cobegin-stacksize.
This test was originally added with the following message:

"Add a test that fails (on linux systems, anyway) unless the new callStackSize
config const is set at execution time, as it is in the .execopts file here."

However the failure mode really depends on the configuration (--baseline,
default, --fast) and exactly what code we're generating. Currently this tests
pases with default without setting CHPL_RT_CALL_STACK_SIZE so it's not really
testing what it intended to anymore (which I think is just that
CHPL_RT_CALL_STACK_SIZE actually changes the call stack size beyond just what
the tasking layer reports to us.)

To make it more portable to different configurations this updates so the test
runs in two modes. One with 32KB which will always fail, and one with 32MB
which should always pass.