Update third-party Makefile to only try re2 once when built speculatively.
Update the third-party/Makefile to only try to build re2 when the build/ subdir
(e.g. build/darwin-clang-native/) is not present (as opposed to checking for
the install dir). The build/ subdir is created early in the process, so even if
re2 fails to compile, it will exist. This allows the third-party to use it as
an indication that the build has been tried (regardless of success or failure).

Update the re2 Makefile to separate build/ subdir creation from regular steps
(and add another variable RE2_BUILD_DIR similar to RE2_INSTALL_DIR). The re2
Makefile still behaves the same as before, but separating out the build subdir
step makes it more clear that it is independent of the re2 build/install step.

All of this means a user that calls make from the top level and has an re2
build failure will not continue trying to re-build re2 (and failing). This
should be a better user experience.