Change signature of chpl_rt_free_c_string to take a c_string_copy instead.
Only the c_string_copy version of the function should be required, but a c_string version
is supplied in the interim as a bridge.  A c_string is "unowned" so it is technically
illegal to free one.

The new string implementation uses c_string as its base type instead of c_string_copy,
because it was created before the distinction was made between "owned" c_string_copy
objects and "unowned" c_string objects.  Rewriting string_rec to use c_string_copy as its
base type is left as a separate task.

TODO: Rewrite string_rec to use c_string_copy as its base type.

###DETAILS
chpl_rt_free_c_string_copy() was rewritten to take its argument by reference and zero out
the referenced c_string_copy pointer when done.  On the Chapel side, the pointer to be
deleted must be passed using ref intent.  In several places, it was necessary to add the
ref intent to an argument or to change "const" to "var" to make a local variable writable.