Bump gasnet timeout again
Even at 15 minutes we still have some timeouts for the small core count
machines. Bumping to a ridiculous 30 minutes for machines with <= 4 cores, and
leaving at 15 minutes for <= 8 cores. If we still have timeouts after this,
I'll cry.

As with the previous commits, this is only until we get real hardware to run on
instead of running oversubscribed.

Also bumping the default for all gasnet runs slightly (not just qthreads with
low core count) because we have occasional timeouts for fifo.