Second update to graph url
This is part of a semi large effort to clean up how we store options in the url
for our graphs. This part adds support for having the suites in the url and
switched from having graphs represented with their  checkbox ids to having
graphs be represented by a normalized title. This makes it easy to have a URL
that is valid for a long period of time. Previously anytime the number of
graphs changed the URL would become outdated. This allows the URL for
individual graphs to be good until any graph title changes (happens
infrequently) and it makes it so we can have URL's for suites. The suites is
useful in hyperlinks that we want around long term such as a link in a blog.

This adds functionality to have the suite in the url and a bunch of helper
routines that were required as part of that and changes individual graphs to be
stored by title. It also adds a bunch of comments and does some cleanup.