Fix paratest.server to respect skipif files.
This was a typo I introduced when adding support for python returning "True".
Note the "my" on line 406, which made that variable local.  Dumb.

Revised the test to be a bit more accommodating:
  Any numerical result greater than zero means "skip this test".
  A text result containing "tRuE" (any combination of UC and LC) will mean "skip this test".

Noticed that tests against COMPOPTS did not succeed, because paratest.server does not
export it (nor EXECOPTS) to the environment.  Did that.