Call setColors for expanded graphs and multi configuration graphs
Set colors is still needed for graphs that will be expanded (to set the dashed
lines) and if multiple configurations are being displayed (also for the dashed
lines.)

Also I changed the graph titles for some of the compiler performance graphs
since they had identical names which screws up detecting which suite is
selected. I should add a check for duplicate titles.