Reduce leaks in module code that uses concatenates c_string_copies.
Since a c_string_copy can be coerced to a c_string, there were many places in the code
where the c_string_copy result of a c_string + c_string concatenation was being used as an
argument to a further concatenation and then being dropped on the floor.

Additional concatenation operators that take explicit c_string_copy
arguments, can bind to and consume these temporary results.  By convention any argument of
the c_string_copy type may be treated as an expiring rvalue so it is legal to consume them
in this way.  (An alternative implementation using rvalue references is a future
direction.)

Most of the changes in this patch are due to the (recently added) restriction that ref
formals cannot bind to rvalues (i.e. unnamed temporaries).  The cure is to give the
temporary a name.

While there, I also cleaned up the interface to stringMove, so it appears to consume its
c_string_copy argument.  This is largely to satisfy the convention, and has no real effect
upon current usage.  However if stringMove were recoded to work like realloc --
conditionally reallocating the buffer as needed to accommodate the source string -- then
the revised interface can be used to advantage to prevent leaks.