Use MAX_LOGICAL with some performance tests
When investigating performance numbers after the most recent release, we
discovered that the NAS Parallel Benchmarks implementations for IS and EP
benefited from setting CHPL_RT_NUM_THREADS_PER_LOCALE to MAX_LOGICAL, and the
implementation for FT also benefited but only with larger problem sizes (size B
and up).  The consensus at our most recent meeting was that we should gather
numbers with these settings where beneficial going forward, so I added this
value to the performance execution options for these tests.