Fix tuple early autoDestroy bug.
In cullExplicitAutoDestroyFlags(), expand the effect to also include "regular" insert auto
destroy flags.

This function removes the autoDestroy flags from values that are copied (through a move)
into the return value variable of a function.  The case that was resolved before dealt
only with explicit autodestroy flags, but since tuples are created through a call to
_build_tuple_always rather than a constructor call (which would be an explicit new), it is
necessary to add the "normal" auto destroy case to the routine.

The original code was written to be very narrow, to avoid surprises.  But testing should
reveal if there are any associated with this small expansion.