Remove reference to /data/sea/* in nightly.
The path /data/sea/* was being used in nightly for syncing performance graphs.
The actual path should have been based off of the log directory. This didn't
cause any issues before because on all performance machines that used nightly
the log dir and the path in /data/sea/* were the same.

This updates the code in nightly that used /data/sea/ and updates the
syncToSourceforge graph that is responsible for rsyncing the graphs to
sourceforge. Sync to sourceforge now takes a path to sync and the destination
to sync it to on sf instead of taking a base path and host and trying to figure
out the path to sync and the dest directory